creighton-prime
Basically, I added some stupid arbitrary/coarbitrary instances for 
MValue and added a bit more boilerplate to the LineAST.
It's not much, but it's what I could get done before going to work this 
morning.  I want to try and get some of the basic ones like 
parseExp done tonight.