interpreter changes
So basically what changes here is that the main function is a little 
smarter.  It's childish of me but I wanted an easy way to quit the 
interpreter that didn't rely on the command "hang" working correctly, so 
now "!q" quits.  If you don't mind I'd like to keep the convention that 
lines that start with a '!' go straight to the interpreter itself.  As 
we add more details to the runtime we can have a whole slew of bang 
commands that dump the symbol table, spit out the parse tree for the 
last line, or anything that'd be good for debugging.

And speaking of debugging, try "i x>y".  It completely ignores the ">y" 
part, and I'm not sure why.  I need to stare at your code a bit more, 
but there's only so much time on a lunch break, y'know?